fixes #133
This change is the result of the output of `yarn --integrity`. Rails' task assets:precompile runs an integrity check when it installs yarn packages and adds them to the yarn.lock file. Without this, the deployments fail because git sees changes to files in the repo at the destination. We need to remember to run `yarn --integrity` instead of just `yarn` in the future. Documentation must be updated.

I am still a bit worried that this will cause problems in the future with killing deployments. If one of the yarn packages gets updated and the lock file isn't updated, then the deployment will bail.