Disable search history (#724)
* Disable search history

This disables the search history feature in Blacklight by making the
Search model readonly. This allows all of Blacklight to keep on working
like nothing's changed, except records are never stored to the database.

A user's session uses are readonly Search record, allowing all of the
current behaviors to function, with minimal overrides to Blacklight. An
additional override to the Search model ensure that any additional or
future attempts to save records will not occur.

* Update mimemagic

Changes to licensing required the gem's maintaners to pull old versions.
This updates us to the latest, legal version.