Use Redis to store searches (#730)
* Use Redis to store searches

Uses Redis to store a user's searches for the duration of their session.

Changes:

* configures Redis for use with the application
* application will still perform without a Redis connection
*  overrides Blacklight::SearchContext to store a user's previous
   searches in Redis for the duration of their session
* removes searches table
* renders Blacklight's search history features unusable
  * users cannot permanently save their searches
  * we don't allow this in the UI anyway

Details:

When performing searches in the catalog, keys such as
"searchhistory:xxx" and "search:yyy" will be written to Redis. The
search history key contains an array of search keys, and a search key
contains the value for a given search stored as a json string.

Keys are expired after a default period of 24 hours, configurable via
environment variables.

If Redis is unavailable, the application will still work, and users can
search, however, pagination through search results will not be
possible.