Browse subject facets (#833)
Adds an endpoint for browsing subject facets. Refactors our existing
browse components so that they are compatible with both call number and
subject browsing features.

This duplicates many of the same features used in browsing facets on the
home page via Blacklight's faceting; however, this does _not_ utilize
any of that as a dependency. Instead, facet browsing is re-implemented
to keep our feature set distinct from Blacklight's so that if anything
changes in the latter, it will not impact us.