Add subject browse to search options
This resurrects a blocked commit for call number browse because we want
to do the same thing with subjects. However, call number browse needs to
be hidden from non-prod hosts, so the option is blocked if we're using
the Matomo ID for production.

Using the same technique we use for call numbers, we add an option for
subject browsing, which redirects to the subject browse form. The
PrefixSelector component is updated to include an input for supplying a
custom prefix other than the canned letters.