Updating our dev-related tasks
This add a new 'dev:clean' task which wipes out everything, loads any
new solr configuration, and loads fixtures. It ensures that your dev
environment is in a clean state and has the latest solr configuration.

Additionally, descriptions are added to the solr.rake tasks so that
they'll list properly when running rake -T.

The method for creating fixtures is moved to a bin executable because
this seemed the best way to navigate the issues with Bundler. Using a
bin command relies on the local system's shell environment.

It's worth noting that we're not using the DatabaseCleaner gem to do the
database truncation because it's not really worth it. Instead of
introducing another dependency, we can truncate the tables via SQLite to
remove the data, and rely on Rails' use_transactional_fixtures to take
care the rest in between tests.