Add options for including lower term (#876)
When retrieving shelf lists with TermsQuery, we only need to include the
lower term, which is the query term or book we're looking for, in either
the forward or reverse set, but not _both_. Including it in both results
in the book present in the forward and reverse set. When the shelf is
then assembled with both queries, the book will appear twice.

Adding an option to include the lower term allows us to only request it
for the reverse set and not the forward set. This creates a virtual
shelf with the queried item only appearing once instead of twice.