Add custom placeholder text for browse queries in search bar
When the user selects one of the browse actions, we now update
the placeholder text to show a helpful hint for that browse type.

We also remember their search/browse type selection if they select
one of the browse options (i.e. we will remember if they selected
Browse by Call Number and not reset the search type dropdown).