Availability A11y: Add paging controls for screen reader users
For easier navigation throughout the availability table as it is
expanded via the "View More" button, we add navigational rows between
each page of holdings. Each row shows the current page you are on,
e.g. "Holdings x - y" and a Previous/Next button to change focus to the
previous or next page of results, respectively.

These controls are hidden for non-screen-reader users, and are only used
for accessibility purposes.