1053 limit search terms (#1303)
* 1053:  Adds clause count filter

Adds a filter to the SearchBuilder processor chain. This filter limits the number
of search terms, depending on the field (ie different number of terms for author than title).

* 1053 - Makes formatting changes requested by Rubocop

* 1053 - no longer applies filter on lucene searches

We only want the parser to work on edismax/dismax queries, not lucene queries, which are used for example with bookmarked pages. They were being truncated in weird ways, and it was not desired.

* 1053 - Adds tests for the clause count limiter

Checks that the number of search terms matches what would be expected.